+++ internal/service/service.go
@@ -3,7 +3,6 @@ package service
 import (
 	"github.com/gin-gonic/gin"
 	"github.com/uselagoon/lagoon/services/insights-handler/internal/lagoonclient"
-	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
 	"net/http"
 	"strconv"
@@ -22,31 +21,6 @@ func SetupRouter(db *gorm.DB) (*gin.Engine, error) {
 	router.POST("/environment/:id/facts", PostFactsByEnvironmentEndpoint)
 	router.DELETE("/environment/:id/facts", DeleteFactsByEnvironmentEndpoint)
 	return router, nil
-}
-
-type Dboptions struct {
-	Filename string
-}
-
-// SetUpDatabase will connect to the selected DB and run pending migrations
-func SetUpDatabase(opts Dboptions) (*gorm.DB, error) {
-	// TODO: currently we're only supporting sqlite for dev
-	// going forward, this will run on mysql - but both should be selected
-
-	db, err := gorm.Open(sqlite.Open(opts.Filename), &gorm.Config{})
-	if err != nil {
-		return db, err
-	}
-
-	if err = db.AutoMigrate(&lagoonclient.Fact{}); err != nil {
-		return db, err
-	}
-
-	if err = db.AutoMigrate(&lagoonclient.LagoonProblem{}); err != nil {
-		return db, err
-	}
-
-	return db, nil
 }
 
 // Fact represents the data model for a fact (assuming it's defined globally)
@@ -163,8 +137,8 @@ func GetFactsByEnvironmentEndpoint(c *gin.Context) {
 	}
 
 	// Query the database for facts by environmentID
-	var facts []lagoonclient.Fact
-	if err := gormDB.Where("environment = ?", environmentID).Find(&facts).Error; err != nil {
+	facts, err := GetFacts(gormDB, environmentID)
+	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to retrieve facts from the database"})
 		return
 	}
